chore(button-timed): remove stateless version of component
